Add utilities for testing collection.
