+++ features/timeseries-evaluate/src/test/java/org/opennms/netmgt/collection/persistence/evaluate/EvaluateStatsIT.java
@@ -33,6 +33,7 @@ import java.util.HashMap;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.opennms.core.collection.test.MockCollectionAgent;
 import org.opennms.core.utils.InetAddressUtils;
 import org.opennms.netmgt.collection.api.AttributeType;
 import org.opennms.netmgt.collection.api.ServiceParameters;
